Migrate test to jest
